+++ actions/core/configure.sh
@@ -22,7 +22,7 @@ core_configure ()
     local configured=false
 
     if [[ -d "$CORE_PATH_CONFIG" ]]; then
-        read -p "We found an Ark Core configuration, do you want to overwrite it? [y/N] : " choice
+        read -p "We found an RipaEx Core configuration, do you want to overwrite it? [y/N] : " choice
 
         if [[ "$choice" =~ ^(yes|y|Y) ]]; then
             __core_configure_pre
@@ -56,7 +56,7 @@ core_configure ()
     fi
 
     if [[ "$configured" = true ]]; then
-        read -p "Ark Core has been configured, would you like to start the relay? [Y/n] : " choice
+        read -p "RipaEx Core has been configured, would you like to start the relay? [Y/n] : " choice
 
         if [[ -z "$choice" || "$choice" =~ ^(yes|y|Y) ]]; then
             relay_start